requirements.txt
python-dotenv
ollama
pyupbit
openai
requests
mysql-connector-python
yfinance
pandas